[BISERVER-13627] rewritten test to remove the potential synchronization deadlock [CLEANUP] moved lost test classes to maven standard location
